Implement inter-process lock test using JVM helper
